Add abc9_ops -reintegrate; moved out from now abc9_exe
